pyproject.toml
[project]
name = "twidilers"
dependencies = [
    "flask",
    "Flask-SQLAlchemy",
    "python-dotenv",
    "psycopg2",
    "flask-bcrypt",
    "pillow",
    "gunicorn",
    "Flask-Mail",
    "requests",
    "PyJWT",
    "ua-parser[regex]",
    "flask-migrate",
    "random-username",
]
version='1.0'
description='A wonderful website'

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"